requirements.txt
fastapi
asyncpg
SQLAlchemy
uvicorn
python-dotenv
pydantic_settings
orjson
jwt
passlib
